Rename sync components
