+++ commonspace/spaceutils_test.go
@@ -18,7 +18,6 @@ import (
 	"github.com/anyproto/any-sync/app/ocache"
 	"github.com/anyproto/any-sync/commonspace/config"
 	"github.com/anyproto/any-sync/commonspace/credentialprovider"
-	"github.com/anyproto/any-sync/commonspace/globalsync"
 	"github.com/anyproto/any-sync/commonspace/object/accountdata"
 	"github.com/anyproto/any-sync/commonspace/object/acl/list"
 	"github.com/anyproto/any-sync/commonspace/object/tree/objecttree"
@@ -44,6 +43,7 @@ import (
 	"github.com/anyproto/any-sync/nodeconf"
 	"github.com/anyproto/any-sync/testutil/accounttest"
 	"github.com/anyproto/any-sync/util/crypto"
+	"github.com/anyproto/any-sync/util/syncqueues"
 )
 
 //
@@ -692,7 +692,7 @@ func newFixtureWithData(t *testing.T, spaceId string, keys *accountdata.AccountK
 		process:              newSpaceProcess(spaceId),
 	}
 	fx.app.Register(fx.account).
-		Register(globalsync.New()).
+		Register(syncqueues.New()).
 		Register(fx.config).
 		Register(peerPool).
 		Register(rpctest.NewTestServer()).
